requirements.txt
fastapi
uvicorn[standard]
requests
pymongo
dnspython
python-dotenv